hw germantitov 1401 question 5.py

+++ hw germantitov 1401 question 5.py
@@ -1,7 +1,8 @@
 width: int= 30
 height: int= 40
 area: float= width * height
-perimeter: int= 2 * (width + height)
+perimeter: float= 2 * (width + height)
 
 print("rectangle area: ", area)
 print("perimeter: ", perimeter)
+
